backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
cycler==0.11.0
dj-database-url==1.0.0
Django==4.1.2
django-cockroachdb==4.1
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
fonttools==4.29.0
gunicorn==20.1.0
idna==3.4
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.4
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
robotframework==6.0
robotframework-requests==0.9.4
scipy==1.7.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
whitenoise==6.2.0
